Fixed revert errors on previous commit.
